requirements-dev.txt
astroid==3.1.0
bandit==1.7.8
dill==0.3.8
flake8==7.0.0
iniconfig==2.0.0
isort==5.13.2
markdown-it-py==3.0.0
mccabe==0.7.0
mdurl==0.1.2
packaging==24.0
pbr==6.0.0
platformdirs==4.2.0
pluggy==1.4.0
pycodestyle==2.11.1
pyflakes==3.2.0
Pygments==2.17.2
pylint==3.1.0
pytest==8.1.1
PyYAML==6.0.1
rich==13.7.1
stevedore==5.2.0
tomlkit==0.12.4
